// source/libs_game/protobuf.d.ts
    declare class Field extends FieldBase {

        /**
         * Constructs a new message field instance. Note that {@link MapField|map fields} have their own class.
         * @param name Unique name within its namespace
         * @param id Unique id within its namespace
         * @param type Value type
         * @param [rule="optional"] Field rule
         * @param [extend] Extended type if different from parent
         * @param [options] Declared options
         */
        constructor(name: string, id: number, type: string, rule?: (string | { [k: string]: any }), extend?: (string | { [k: string]: any }), options?: { [k: string]: any });

        /**
         * Constructs a field from a field descriptor.
         * @param name Field name
         * @param json Field descriptor
         * @returns Created field
         * @throws {TypeError} If arguments are invalid
         */
        public static fromJSON(name: string, json: IField): Field;

        /** Determines whether this field is packed. Only relevant when repeated and working with proto2. */
        public readonly packed: boolean;

        /**
         * Field decorator (TypeScript).
         * @param fieldId Field id
         * @param fieldType Field type
         * @param [fieldRule="optional"] Field rule
         * @param [defaultValue] Default value
         * @returns Decorator function
         */
        public static d<T extends number | number[] | Long | Long[] | string | string[] | boolean | boolean[] | Uint8Array | Uint8Array[] | Buffer | Buffer[]>(fieldId: number, fieldType: ("double" | "float" | "int32" | "uint32" | "sint32" | "fixed32" | "sfixed32" | "int64" | "uint64" | "sint64" | "fixed64" | "sfixed64" | "string" | "bool" | "bytes" | object), fieldRule?: ("optional" | "required" | "repeated"), defaultValue?: T): FieldDecorator;

        /**
         * Field decorator (TypeScript).
         * @param fieldId Field id
         * @param fieldType Field type
         * @param [fieldRule="optional"] Field rule
         * @returns Decorator function
         */
        public static d<T extends Message<T>>(fieldId: number, fieldType: (Constructor<T> | string), fieldRule?: ("optional" | "required" | "repeated")): FieldDecorator;
    }

    /** Base class of all reflected message fields. This is not an actual class but here for the sake of having consistent type definitions. */
    declare class FieldBase extends ReflectionObject {

        /**
         * Not an actual constructor. Use {@link Field} instead.
         * @param name Unique name within its namespace
         * @param id Unique id within its namespace
         * @param type Value type
         * @param [rule="optional"] Field rule
         * @param [extend] Extended type if different from parent
         * @param [options] Declared options
         * @param [comment] Comment associated with this field
         */
        constructor(name: string, id: number, type: string, rule?: (string | { [k: string]: any }), extend?: (string | { [k: string]: any }), options?: { [k: string]: any }, comment?: string);

        /** Field rule, if any. */
        public rule?: string;

        /** Field type. */
        public type: string;

        /** Unique field id. */
        public id: number;

        /** Extended type if different from parent. */
        public extend?: string;

        /** Whether this field is required. */
        public required: boolean;

        /** Whether this field is optional. */
        public optional: boolean;

        /** Whether this field is repeated. */
        public repeated: boolean;

        /** Whether this field is a map or not. */
        public map: boolean;

        /** Message this field belongs to. */
        public message: (Type | null);

        /** OneOf this field belongs to, if any, */
        public partOf: (OneOf | null);

        /** The field type's default value. */
        public typeDefault: any;

        /** The field's default value on prototypes. */
        public defaultValue: any;

        /** Whether this field's value should be treated as a long. */
        public long: boolean;

        /** Whether this field's value is a buffer. */
        public bytes: boolean;

        /** Resolved type if not a basic type. */
        public resolvedType: (Type | Enum | null);

        /** Sister-field within the extended type if a declaring extension field. */
        public extensionField: (Field | null);

        /** Sister-field within the declaring namespace if an extended field. */
        public declaringField: (Field | null);

        /** Comment for this field. */
        public comment: (string | null);

        /**
         * Converts this field to a field descriptor.
         * @param [toJSONOptions] JSON conversion options
         * @returns Field descriptor
         */
        public toJSON(toJSONOptions?: IToJSONOptions): IField;

        /**
         * Resolves this field's type references.
         * @returns `this`
         * @throws {Error} If any reference cannot be resolved
         */
        public resolve(): Field;
    }

    /** Field descriptor. */
    declare interface IField {

        /** Field rule */
        rule?: string;

        /** Field type */
        type: string;

        /** Field id */
        id: number;

        /** Field options */
        options?: { [k: string]: any };
    }

    /** Extension field descriptor. */
    declare interface IExtensionField extends IField {

        /** Extended type */
        extend: string;
    }

    /**
     * Decorator function as returned by {@link Field.d} and {@link MapField.d} (TypeScript).
     * @param prototype Target prototype
     * @param fieldName Field name
     */
    type FieldDecorator = (prototype: object, fieldName: string) => void;

    /**
     * A node-style callback as used by {@link load} and {@link Root#load}.
     * @param error Error, if any, otherwise `null`
     * @param [root] Root, if there hasn't been an error
     */
    type LoadCallback = (error: (Error | null), root?: Root) => void;
// ...
    declare interface INamespace {

        /** Namespace options */
        options?: { [k: string]: any };

        /** Nested object descriptors */
        nested?: { [k: string]: AnyNestedObject };
    }

    /** Any extension field descriptor. */
    type AnyExtensionField = (IExtensionField | IExtensionMapField);

    /** Any nested object descriptor. */
    type AnyNestedObject = (IEnum | IType | IService | AnyExtensionField | INamespace);
// ...
    declare interface IParseOptions {

        /** Keeps field casing instead of converting to camel case */
        keepCase?: boolean;

        /** Recognize double-slash comments in addition to doc-block comments. */
        alternateCommentMode?: boolean;
    }
// ...
    declare class Root extends NamespaceBase {

        /**
         * Constructs a new root namespace instance.
         * @param [options] Top level options
         */
        constructor(options?: { [k: string]: any });

        /** Deferred extension fields. */
        public deferred: Field[];

        /** Resolved file names of loaded files. */
        public files: string[];

        /**
         * Loads a namespace descriptor into a root namespace.
         * @param json Nameespace descriptor
         * @param [root] Root namespace, defaults to create a new one if omitted
         * @returns Root namespace
         */
        public static fromJSON(json: INamespace, root?: Root): Root;

        /**
         * Resolves the path of an imported file, relative to the importing origin.
         * This method exists so you can override it with your own logic in case your imports are scattered over multiple directories.
         * @param origin The file name of the importing file
         * @param target The file name being imported
         * @returns Resolved path to `target` or `null` to skip the file
         */
        public resolvePath(origin: string, target: string): (string | null);

        /**
         * Loads one or multiple .proto or preprocessed .json files into this root namespace and calls the callback.
         * @param filename Names of one or multiple files to load
         * @param options Parse options
         * @param callback Callback function
         */
        public load(filename: (string | string[]), options: IParseOptions, callback: LoadCallback): void;

        /**
         * Loads one or multiple .proto or preprocessed .json files into this root namespace and calls the callback.
         * @param filename Names of one or multiple files to load
         * @param callback Callback function
         */
        public load(filename: (string | string[]), callback: LoadCallback): void;

        /**
         * Loads one or multiple .proto or preprocessed .json files into this root namespace and returns a promise.
         * @param filename Names of one or multiple files to load
         * @param [options] Parse options. Defaults to {@link parse.defaults} when omitted.
         * @returns Promise
         */
        public load(filename: (string | string[]), options?: IParseOptions): Promise<Root>;

        /**
         * Synchronously loads one or multiple .proto or preprocessed .json files into this root namespace (node only).
         * @param filename Names of one or multiple files to load
         * @param [options] Parse options. Defaults to {@link parse.defaults} when omitted.
         * @returns Root namespace
         * @throws {Error} If synchronous fetching is not supported (i.e. in browsers) or if a file's syntax is invalid
         */
        public loadSync(filename: (string | string[]), options?: IParseOptions): Root;
    }
// ...
    declare interface Constructor<T> extends Function {
        new(...params: any[]): T; prototype: T;
    }

    /** Properties type. */
    type Properties<T> = { [P in keyof T]?: T[P] };

    /**
     * Any compatible Buffer instance.
     * This is a minimal stand-alone definition of a Buffer instance. The actual type is that declareed by node's typings.
     */
    declare interface Buffer extends Uint8Array {
    }

    /**
     * Any compatible Long instance.
     * This is a minimal stand-alone definition of a Long instance. The actual type is that declareed by long.js.
     */
    declare interface Long {

        /** Low bits */
        low: number;

        /** High bits */
        high: number;

        /** Whether unsigned or not */
        unsigned: boolean;
    }